requirements-minimal.txt
# Minimal dependencies for Teams Bot only

# Environment variables
python-dotenv==1.0.0

# Azure DevOps
azure-devops==7.1.0b4
msrest==0.7.1

# Microsoft Teams Bot
botbuilder-core==4.15.0
aiohttp==3.10.11
